Feature: Add a route to register the login of a user and a route to register the logout
