+++ src/main/java/br/com/client/micro/exceptions/PermissionDeniedException.java
@@ -1,0 +1,8 @@
+package br.com.client.micro.exceptions;
+
+public class PermissionDeniedException extends RuntimeException {
+    public PermissionDeniedException(){super("You do not have permission to perform this action!");}
+    public PermissionDeniedException(String message) {
+        super(message);
+    }
+}
