+++ src/main/java/br/com/client/micro/repository/IClientRepository.java
@@ -6,9 +6,13 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface IClientRepository extends MongoRepository<Client, String> {
     public boolean existsByCpf(String cpf);
     public boolean existsByEmail(String email);
     Page<Client> findByRole_Name(String roleName, Pageable pageable);
+    public Optional<Client> findByEmail(String email);
+
 }
